+++ codex-rs/default.nix
@@ -2,6 +2,7 @@
   cmake,
   llvmPackages,
   openssl,
+  fetchurl,
   libcap ? null,
   rustPlatform,
   pkg-config,
@@ -14,9 +15,21 @@ rustPlatform.buildRustPackage (_: {
   env.PKG_CONFIG_PATH = lib.makeSearchPathOutput "dev" "lib/pkgconfig" (
     [ openssl ] ++ lib.optionals stdenv.isLinux [ libcap ]
   );
+  env.CARGO_PROFILE_RELEASE_LTO = "false";
+  env.RUSTY_V8_ARCHIVE = fetchurl {
+    name = "librusty_v8-146.4.0";
+    url = "https://github.com/denoland/rusty_v8/releases/download/v146.4.0/librusty_v8_release_${stdenv.hostPlatform.rust.rustcTarget}.a.gz";
+    hash = "sha256-5ktNmeSuKTouhGJEqJuAF4uhA4LBP7WRwfppaPUpEVM=";
+  };
   pname = "codex-rs";
   inherit version;
   cargoLock.lockFile = ./Cargo.lock;
+  cargoBuildFlags = [
+    "-p"
+    "codex-cli"
+    "--bin"
+    "codex"
+  ];
   doCheck = false;
   src = ./.;
 
@@ -43,6 +56,7 @@ rustPlatform.buildRustPackage (_: {
     "nucleo-0.5.0" = "sha256-Hm4SxtTSBrcWpXrtSqeO0TACbUxq3gizg1zD/6Yw/sI=";
     "nucleo-matcher-0.3.1" = "sha256-Hm4SxtTSBrcWpXrtSqeO0TACbUxq3gizg1zD/6Yw/sI=";
     "runfiles-0.1.0" = "sha256-uJpVLcQh8wWZA3GPv9D8Nt43EOirajfDJ7eq/FB+tek=";
+    "libwebrtc-0.3.26" = "sha256-0HPuwaGcqpuG+Pp6z79bCuDu/DyE858VZSYr3DKZD9o=";
     "tokio-tungstenite-0.28.0" = "sha256-hJAkvWxDjB9A9GqansahWhTmj/ekcelslLUTtwqI7lw=";
     "tungstenite-0.27.0" = "sha256-AN5wql2X2yJnQ7lnDxpljNw0Jua40GtmT+w3wjER010=";
   };